🎁: Implemented Report access information (#585)
* Modified Report error component to show error details

Updated exisitng error component to show report and access information
when the user does not have access to the report.
Added avatar for the person responsible for the report.

Added new prop for pbi reports, that provides the dialog with
information.

* Prettified Code!

* Updated Power BI Package to be fusion-independent

Removed deps to fusion libraries
Implemented fallback component handler to report error boundary.

* Removed old error (fusion-dependent) components.

* Bumped package.json

---------

Co-authored-by: kjellhaaland <kjellhaaland@users.noreply.github.com>